DS-5048 by chmez: Improve process for building list of thread members and show special text instead of a message of deleted user.
