+++ modules/social_features/social_private_message/config/install/core.entity_view_display.private_message_thread.private_message_thread.default.yml
@@ -16,7 +16,7 @@ content:
     third_party_settings: {  }
   members:
     label: hidden
-    type: private_message_thread_member_formatter
+    type: social_private_message_thread_member_formatter
     weight: 0
     region: content
     settings:
